Removal of redundant files and renaming of some files.
