+++ yolo2ncnn.py
@@ -1,0 +1,5 @@
+from ultralytics import YOLO
+
+model = YOLO("yolov8n.pt")
+
+model.export(format = "ncnn")
